backend/requirements.txt
numpy
flask
flask_cors
matplotlib
opencv-python
#PyQt5
paho-mqtt
wlkata-mirobot-python